Introduced Non-Linear Pose Estimation Equations
Updated the getX and getY methods in ForwardKinematic to utilize some fancier equations that now model the robot as travelling in an arc between two timesteps (more accurate than the earlier assumption that the drivetrain was travelling in a straight line between timesteps). In addition, the heading methods have been deprecated to prepare for future removal from the class (calculating heading information based on odometery is useless when much more accurate orientation data can be provided by the gyroscopic sensor onboard the robot (in the IMU)).